backend/pyproject.toml
[project]
name = "raw-marketplace"
version = "0.1.0"
description = "Backend side of Niff Marketplace project based on Django Rest Framework"
authors = [
    {name = "kabust",email = "tulyulyuk91@gmail.com"}
]
license = {text = "Apache 2.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-debug-toolbar (>=5.1.0,<6.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "flake8-pyproject (>=1.2.3,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "markdown (>=3.8,<4.0)",
    "django-filter (>=25.1,<26.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-oauth-toolkit (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.flake8]
max-line-length = 99
exclude = [
    ".git",
    "__pycache__",
    "venv",
    "migrations"
]
extend-ignore = [
    "E203"
]

[tool.black]
line-length = 99
target-version = ['py313']
extend-exclude = '''
(
  ^/venv/
)
'''
